requirements.txt
flask
bcrypt
pandas
openpyxl
selenium
flask-mysqldb
python-dotenv
flask_sqlalchemy
webdriver-manager
flask-jwt-extended